refactor(preview): give the zoom overlay its own file and register its icons
The preview is about to grow a diff renderer, diagram captions and a hover
toolbar. DiagramZoom already owns a full-screen surface with its own key
handling, so leaving it inside MarkdownPreview.tsx would make that file the
place where every unrelated concern lands.

The five icons the upcoming toolbar and zoom controls need are registered
now, while nothing else is moving, so the behavioural commits that follow
touch only their own logic. copy/check were already imported raw by the
editor node view; only copy was missing from the table.

No behaviour change.
